Beta 4 (#4)
* Complete change in the API(Not working)

An error is always raised.

* Delete BotLangEN.py

* Language changes

DE, PT, ES and RU coming soon

* Main bot updated

Many new updates coming to work with the new API

* Updated Readme.md with all the info it was needed

Raw note: API is not working, I'm making various bug fixes.

* Added __init__.py and Config.py

The __init__.py serves to give info about the creator, and Config.py is used to set the main things of the bot.

* Update Config.py

* Updated Bot.py - Now everything works!

Now all commands are working as expected! Also updated to work with the changes in the Language and API

* Updated EN.py to work with the Config.py

Now you must set up the configurations in Config.py :D

* Complete change in Img.py

Now the API is fully new, and working properly o/

* Update LICENSE

* Updated Readme.md with all the info it was needed

* Updated Readme.md with all the info it was needed

* Updated /help and added new features

New error logs, in case something goes wrong, and new try/except in the commands to avoid infinite loop.

* New strings added.

New strings added to work with the update in Bot.py

* Finally started working in Inline.

Inline is not working yet, I'm trying.

* New function Query, for Inline mode.

The func Query(tag) searches for the respective tag.

* Uploaded acc_setup.py

This file configs the username and password to connect with the repository.

* Correct mispelling

Now it's all ok, I think.

* Updated README with missing stuff

Now everything that must be installed is correctly presented.

* Added database folder

Saving the ID's for future options.

* Finally started working in Inline.

Inline is not working yet, I'm trying.