feat(cloud_handler): :rocket: introduce cloud-ready codebase
This commit brings the first iteration for the serverless context the
bot will live in. It introduces handlers for receiving messages from
Cloud Pub/Sub, and local long polling for faster iterations.
We are also introducing a new way of working with bot message
handlers, that are set up with dependency injection (must look
after this further).
Also introduce all-file copyright claims.